Add files via upload
Dear Team.
I am still working on Building Model Part. But if any of you working on the final document and would like to incorporate any of the data exploration, data preparation from this please do it. I will update this as soon as I finished the last part of building models. 
Thank You.
Don Padmaperuma